Fix interface management for freebsd and darwin
